+++ flutter_vlc_player/example/lib/video_data.dart
@@ -1,11 +1,3 @@
-// ignore: prefer-match-file-name
-enum VideoType {
-  asset,
-  file,
-  network,
-  recorded,
-}
-
 class VideoData {
   final String name;
   final String path;
@@ -16,4 +8,11 @@ class VideoData {
     this.path,
     this.type,
   });
+}
+
+enum VideoType {
+  asset,
+  file,
+  network,
+  recorded,
 }
