+++ flutter_vlc_player/example/test/widget_test.dart
@@ -7,8 +7,7 @@
 
 import 'package:flutter/material.dart';
 import 'package:flutter_test/flutter_test.dart';
-
-import 'package:flutter_vlc_player_example/main.dart';
+import 'package:flutter_vlc_player_example/app.dart';
 
 void main() {
   testWidgets('Verify Platform version', (WidgetTester tester) async {
