+++ flutter_vlc_player_platform_interface/lib/src/enums/vlc_media_event_type.dart
@@ -1,4 +1,3 @@
-// ignore: prefer-match-file-name
 enum VlcMediaEventType {
   opening,
   playing,
