+++ flutter_vlc_player_platform_interface/lib/src/events/vlc_media_event.dart
@@ -1,8 +1,7 @@
 import 'dart:ui';
 
-import 'package:flutter_vlc_player_platform_interface/src/enums/media_event_type.dart';
+import 'package:flutter_vlc_player_platform_interface/src/enums/vlc_media_event_type.dart';
 
-// ignore: prefer-match-file-name
 class VlcMediaEvent {
   /// The type of the event.
   final VlcMediaEventType mediaEventType;
