+++ flutter_vlc_player_platform_interface/lib/src/events/vlc_renderer_event.dart
@@ -1,6 +1,5 @@
-import 'package:flutter_vlc_player_platform_interface/src/enums/renderer_event_type.dart';
+import 'package:flutter_vlc_player_platform_interface/src/enums/vlc_renderer_event_type.dart';
 
-// ignore: prefer-match-file-name
 class VlcRendererEvent {
   /// The type of the event.
   final VlcRendererEventType eventType;
